Fix darkened camera view
The recent Three.js relies on WebGL for sRGB - Linear color
space conversion. But it does in shader for video textures
for a performance reason.

camera-tool sets isVideoTexture = true for the screen render
target texture to update the render target only when the screens
are in view as hack. But this hack causes a wrong color space
conversion and the camera views are darkened.

This commit resolves the problem by removing the hack and
rewriting it more properly.